+++ client/package.json
@@ -78,6 +78,7 @@
     "start-js": "BROWSER=none REACT_APP_UI_DEV_MODE=$UI_DEV_MODE REACT_APP_RENKU_UI_URL=$RENKU_UI_URL craco --max_old_space_size=4096 start",
     "start": "react-scripts start",
     "build": "REACT_APP_UI_DEV_MODE=false craco --max_old_space_size=4096 build",
+    "build:analyze": "REACT_APP_UI_DEV_MODE=false craco --max_old_space_size=4096 build --analyze",
     "test": "craco test",
     "eject": "craco eject",
     "lint": "eslint --max-warnings=0 src/",
