+++ client/src/components/form-field/LazyCkEditor.tsx
@@ -19,16 +19,14 @@
 import { ComponentProps, Suspense, lazy } from "react";
 import { Loader } from "../Loader";
 
-const NotebookRender = lazy(() => import("./NotebookRender"));
+const CkEditor = lazy(() => import("./CkEditor"));
 
-// ? Lazy loading of NotebookRender allows us to split off ~950kB from
+// ? Lazy loading of CkEditor allows us to split off ~270kB from
 // ? the main bundle.
-export default function LazyNotebookRender(
-  props: ComponentProps<typeof NotebookRender>
-) {
+export default function LazyCkEditor(props: ComponentProps<typeof CkEditor>) {
   return (
     <Suspense fallback={<Loader />}>
-      <NotebookRender {...props} />
+      <CkEditor {...props} />
     </Suspense>
   );
 }
