+++ client/src/components/notebook/NotebookRender.tsx
