+++ client/src/file/File.present.js
@@ -16,11 +16,11 @@
  * limitations under the License.
  */
 
-import React, { memo, useState } from "react";
 import { faGitlab } from "@fortawesome/free-brands-svg-icons";
 import { faDownload } from "@fortawesome/free-solid-svg-icons";
 import DOMPurify from "dompurify";
 import _ from "lodash";
+import React, { memo, useState } from "react";
 import {
   Badge,
   Button,
@@ -37,13 +37,13 @@ import { Clipboard } from "../components/Clipboard";
 import { ExternalIconLink, ExternalLink } from "../components/ExternalLinks";
 import { Loader } from "../components/Loader";
 import { TimeCaption } from "../components/TimeCaption";
-// import NotebookPreview from "../components/notebook";
 import LazyNotebookPreview from "../components/notebook/LazyNotebookRender";
 import { CheckNotebookStatus } from "../notebooks";
 import { CheckNotebookIcon } from "../notebooks/NotebookStart.present";
 import { formatBytes } from "../utils/helpers/HelperFunctions";
 import { FileAndLineageSwitch } from "./FileAndLineageComponents";
 import { FilePreview } from "./index";
+
 import "../../node_modules/highlight.js/styles/atom-one-light.css";
 
 const commitMessageLengthLimit = 120;
@@ -501,23 +501,11 @@ const StyledNotebook = memo((props) => {
       setDisplayMode={setDisplayMode}
     />,
     <CardBody key="notebook">
-      {/* <Suspense fallback={<Loader />}>
-        <LazyNotebookPreview
-          defaultStyle={false}
-          loadMathjax={false}
-          notebook={notebook}
-        />
-      </Suspense> */}
       <LazyNotebookPreview
         defaultStyle={false}
         loadMathjax={false}
         notebook={notebook}
       />
-      {/* <NotebookPreview
-        defaultStyle={false}
-        loadMathjax={false}
-        notebook={notebook}
-      /> */}
     </CardBody>,
   ];
 }, _.isEqual);
