+++ client/src/file/FilePreview.tsx
@@ -16,18 +16,17 @@
  * limitations under the License.
  */
 
-import hljs from "highlight.js";
 import React from "react";
-import { CardBody } from "reactstrap";
 import { Document, Page } from "react-pdf/dist/esm/entry.webpack";
-import DOMPurify from "dompurify";
+import { CardBody } from "reactstrap";
+
+import { RenkuMarkdown } from "../components/markdown/RenkuMarkdown";
+import { encodeImageBase64 } from "../components/markdown/RenkuMarkdownWithPathTranslation";
+import { atobUTF8 } from "../utils/helpers/Encoding";
+import { FileNoPreview, StyledNotebook } from "./File.present";
+import LazyCodePreview from "./LazyCodePreview";
 
 import "react-pdf/dist/esm/Page/AnnotationLayer.css";
-
-import { FileNoPreview, StyledNotebook } from "./File.present";
-import { atobUTF8 } from "../utils/helpers/Encoding";
-import { encodeImageBase64 } from "../components/markdown/RenkuMarkdownWithPathTranslation";
-import { RenkuMarkdown } from "../components/markdown/RenkuMarkdown";
 
 /* eslint-disable spellcheck/spell-checker */
 const IMAGE_EXTENSIONS = ["jpg", "jpeg", "png", "tiff", "gif", "svg"];
@@ -252,7 +251,7 @@ function FilePreview(props: FilePreviewProps) {
   if (fileIsCode) {
     return (
       <CardBody key="file preview" className="pb-0 bg-white">
-        <CodePreview
+        <LazyCodePreview
           content={props.file.content}
           // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
           fileExtension={getFileExtension()!}
@@ -277,49 +276,6 @@ function FilePreview(props: FilePreviewProps) {
     <CardBody key="file preview" className="pb-0 bg-white">
       <p>{`Unable to preview file with extension .${getFileExtension()}`}</p>
     </CardBody>
-  );
-}
-
-/* eslint-disable */
-// See https://github.com/highlightjs/highlight.js/blob/main/SUPPORTED_LANGUAGES.md
-const hljsNameMap: Record<string, string> = {
-  jl: "julia",
-  f: "fortran",
-  for: "fortran",
-  ftn: "fortran",
-  fpp: "fortran",
-  f03: "fortran",
-  f08: "fortran",
-  m: "objectivec",
-  mat: "matlab",
-};
-/* eslint-enable */
-function extensionToHljsName(ext: string) {
-  return hljsNameMap[ext] ?? ext;
-}
-
-type CodePreviewProps = {
-  content: string;
-  fileExtension: string;
-};
-
-function CodePreview(props: CodePreviewProps) {
-  const codeBlock = React.useRef<HTMLPreElement>(null);
-  React.useEffect(() => {
-    if (codeBlock.current) {
-      codeBlock.current.innerHTML = DOMPurify.sanitize(
-        codeBlock.current.innerHTML
-      );
-      hljs.highlightBlock(codeBlock.current);
-    }
-  }, [codeBlock]);
-
-  const languageName = extensionToHljsName(props.fileExtension);
-
-  return (
-    <pre ref={codeBlock} className={`hljs language-${languageName} bg-white`}>
-      {atobUTF8(props.content)}
-    </pre>
   );
 }
